refactor : 코드 리팩터링 및 UI 수정
